commit from linux with needles updated

+++ montecarlo_needlesdouble.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 double pi = 3.14159265;
 
-bool check_intersect(float x, float theta, int l, int s){
+bool check_intersect(double x, double theta, int l, int s){
     //cout << x << endl;
     //cout << theta << endl;
     if(x + (s/2.0)*sin(theta) >= l)return true;
@@ -20,8 +20,8 @@ int monte_carlo(int t, int l, int s){
     mt19937 gen(rd());
     uniform_real_distribution<> dis(0.0, 1.0);
     int i;
-    float x0;
-    float theta;
+    double x0;
+    double theta;
     for(i = 0; i < t; i++){       
         x0 = l*dis(gen);
         theta = pi*dis(gen);
@@ -34,7 +34,7 @@ int monte_carlo(int t, int l, int s){
 
 int main(int argc, char *argv[]){
     int n, t, l, s, rank, size, pi2, pi2total=0, rc;
-    float t1, t2, tf;   // number of throws, dist between lines and 
+    double t1, t2, tf;   // number of throws, dist between lines and 
     t = (argc > 1 ? atoi(argv[1]): 100000);
     l = (argc > 2 ? atoi(argv[2]): 30);
     s = (argc > 3 ? atoi(argv[3]): 10);
@@ -56,7 +56,7 @@ int main(int argc, char *argv[]){
     t2=MPI_Wtime();
 
     if(rank==0){
-	float pi3 = (2.0*s*t)/(l*float(pi2total));
+	double pi3 = (2.0*s*t)/(l*double(pi2total));
         tf=t2-t1;
         cout << t <<";"<< fixed << setprecision(10) << pi3 << ";" << tf << endl;
     }
